improve request id datatype handling
requests with an id like "/customer(123)/orders" are now checked
concerning the datatype of the id field of the belonging model. So for
example, if a request is made with an id of type "decimal" where as the
id of the model is of type "string", the request is rejected.